django_project/danadjango/requirements.txt
# drf-yasg still imports pkg_resources; setuptools 82+ dropped it
setuptools>=69,<82

Django==4.2.28
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.2.0
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-filter==23.5

# OCR libraries (install manually if needed)
# pytesseract==0.3.10
# opencv-python-headless==4.8.1.78

# Google Vision API (optional, for better OCR)
# google-cloud-vision==3.4.4
# google-auth==2.23.4
